requirements.txt
pydantic==2.11.5
fastapi==0.115.12
uvicorn==0.32.0
groq==0.25.0
python-dotenv==1.1.0
langchain-community
langchain-openai
langchain-groq
openai
faiss-cpu
jq

